changed getmechanisms function
Removed the opening and closing of the selenium browser from the function
scope. This means that opening must be done prior to calling the function
and closing must be done after calling the function. The function also needs
the driver as a parameter - I figured this would make the code cleaner and
calls atention to the fact that the browser must be opened before the function
is called.

This speeds up the code by a factor of almost 3, since now the browser doesn't
need to reopen and close for each url, and instead just opens a new address
on the current window.